+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -1512,10 +1512,53 @@ describe('moduleFileExtensions', () => {
   });
 });
 
+describe('cwd', () => {
+  it('is set to process.cwd', () => {
+    const {options} = normalize({rootDir: '/root/'}, {});
+    expect(options.cwd).toBe(process.cwd());
+  });
+
+  it('is not lost if the config has its own cwd property', () => {
+    console.warn.mockImplementation(() => {});
+    const {options} = normalize(
+      {
+        rootDir: '/root/',
+        cwd: '/tmp/config-sets-cwd-itself',
+      },
+      {},
+    );
+    expect(options.cwd).toBe(process.cwd());
+    expect(console.warn).toHaveBeenCalled();
+  });
+});
+
 describe('Defaults', () => {
   it('should be accepted by normalize', () => {
     normalize({...Defaults, rootDir: '/root'}, {});
 
     expect(console.warn).not.toHaveBeenCalled();
   });
+});
+
+describe('displayName', () => {
+  test.each`
+    displayName             | description
+    ${{}}                   | ${'is an empty object'}
+    ${{name: 'hello'}}      | ${'missing color'}
+    ${{color: 'green'}}     | ${'missing name'}
+    ${{color: 2, name: []}} | ${'using invalid values'}
+  `(
+    'should throw an error when displayName is $description',
+    ({displayName}) => {
+      expect(() => {
+        normalize(
+          {
+            rootDir: '/root/',
+            displayName,
+          },
+          {},
+        );
+      }).toThrowErrorMatchingSnapshot();
+    },
+  );
 });