+++ packages/jest-snapshot/src/inline_snapshots.ts
@@ -50,6 +50,7 @@ const saveSnapshotsForFile = (
 ) => {
   const sourceFile = fs.readFileSync(sourceFilePath, 'utf8');
 
+<<<<<<< HEAD
   let newSourceFile;
   if (prettier) {
     // Resolve project configuration.
@@ -105,9 +106,41 @@ const saveSnapshotsForFile = (
       );
     }, sourceFile);
   }
+=======
+  // Resolve project configuration.
+  // For older versions of Prettier, do not load configuration.
+  const config = prettier.resolveConfig
+    ? prettier.resolveConfig.sync(sourceFilePath, {
+        editorconfig: true,
+      })
+    : null;
 
-  if (newSourceFile !== sourceFile) {
-    fs.writeFileSync(sourceFilePath, newSourceFile);
+  // Detect the parser for the test file.
+  // For older versions of Prettier, fallback to a simple parser detection.
+  const inferredParser = prettier.getFileInfo
+    ? prettier.getFileInfo.sync(sourceFilePath).inferredParser
+    : (config && config.parser) || simpleDetectParser(sourceFilePath);
+
+  // Insert snapshots using the custom parser API. After insertion, the code is
+  // formatted, except snapshot indentation. Snapshots cannot be formatted until
+  // after the initial format because we don't know where the call expression
+  // will be placed (specifically its indentation).
+  const newSourceFile = prettier.format(sourceFile, {
+    ...config,
+    filepath: sourceFilePath,
+    parser: createInsertionParser(snapshots, inferredParser, babelTraverse),
+  });
+>>>>>>> master
+
+  // Format the snapshots using the custom parser API.
+  const formattedNewSourceFile = prettier.format(newSourceFile, {
+    ...config,
+    filepath: sourceFilePath,
+    parser: createFormattingParser(inferredParser, babelTraverse),
+  });
+
+  if (formattedNewSourceFile !== sourceFile) {
+    fs.writeFileSync(sourceFilePath, formattedNewSourceFile);
   }
 };
 
@@ -129,7 +162,45 @@ const groupSnapshotsByFrame = groupSnapshotsBy(({frame: {line, column}}) =>
 );
 const groupSnapshotsByFile = groupSnapshotsBy(({frame: {file}}) => file);
 
-const createParser = (
+const indent = (snapshot: string, numIndents: number, indentation: string) => {
+  const lines = snapshot.split('\n');
+  return lines
+    .map((line, index) => {
+      if (index === 0) {
+        // First line is either a 1-line snapshot or a blank line.
+        return line;
+      } else if (index !== lines.length - 1) {
+        // Do not indent empty lines.
+        if (line === '') {
+          return line;
+        }
+
+        // Not last line, indent one level deeper than expect call.
+        return indentation.repeat(numIndents + 1) + line;
+      } else {
+        // The last line should be placed on the same level as the expect call.
+        return indentation.repeat(numIndents) + line;
+      }
+    })
+    .join('\n');
+};
+
+const getAst = (
+  parsers: {[key: string]: (text: string) => any},
+  inferredParser: string,
+  text: string,
+) => {
+  // Flow uses a 'Program' parent node, babel expects a 'File'.
+  let ast = parsers[inferredParser](text);
+  if (ast.type !== 'File') {
+    ast = file(ast, ast.comments, ast.tokens);
+    delete ast.program.comments;
+  }
+  return ast;
+};
+
+// This parser inserts snapshots into the AST.
+const createInsertionParser = (
   snapshots: Array<InlineSnapshot>,
   inferredParser: string,
   babelTraverse: BabelTraverse,
@@ -142,6 +213,7 @@ const createParser = (
   options.parser = inferredParser;
   const ast = parsers[inferredParser](text);
 
+<<<<<<< HEAD
   traverseAst(snapshots, ast, babelTraverse);
 
   return ast;
@@ -161,6 +233,12 @@ const traverseAst = (
   const groupedSnapshots = groupSnapshotsByFrame(snapshots);
   const remainingSnapshots = new Set(snapshots.map(({snapshot}) => snapshot));
 
+=======
+  const groupedSnapshots = groupSnapshotsByFrame(snapshots);
+  const remainingSnapshots = new Set(snapshots.map(({snapshot}) => snapshot));
+
+  const ast = getAst(parsers, inferredParser, text);
+>>>>>>> master
   babelTraverse(ast, {
     CallExpression({node}) {
       const {arguments: args, callee} = node;
@@ -206,6 +284,70 @@ const traverseAst = (
   if (remainingSnapshots.size) {
     throw new Error(`Jest: Couldn't locate all inline snapshots.`);
   }
+};
+
+// This parser formats snapshots to the correct indentation.
+const createFormattingParser = (
+  inferredParser: string,
+  babelTraverse: Function,
+) => (
+  text: string,
+  parsers: {[key: string]: (text: string) => any},
+  options: any,
+) => {
+  // Workaround for https://github.com/prettier/prettier/issues/3150
+  options.parser = inferredParser;
+
+  const ast = getAst(parsers, inferredParser, text);
+  babelTraverse(ast, {
+    CallExpression({node: {arguments: args, callee}}: {node: CallExpression}) {
+      if (
+        callee.type !== 'MemberExpression' ||
+        callee.property.type !== 'Identifier' ||
+        callee.property.name !== 'toMatchInlineSnapshot' ||
+        !callee.loc ||
+        callee.computed
+      ) {
+        return;
+      }
+
+      let snapshotIndex: number | undefined;
+      let snapshot: string | undefined;
+      for (let i = 0; i < args.length; i++) {
+        const node = args[i];
+        if (node.type === 'TemplateLiteral') {
+          snapshotIndex = i;
+          snapshot = node.quasis[0].value.raw;
+        }
+      }
+      if (snapshot === undefined || snapshotIndex === undefined) {
+        return;
+      }
+
+      const useSpaces = !options.useTabs;
+      snapshot = indent(
+        snapshot,
+        Math.ceil(
+          useSpaces
+            ? callee.loc.start.column / options.tabWidth
+            : callee.loc.start.column / 2, // Each tab is 2 characters.
+        ),
+        useSpaces ? ' '.repeat(options.tabWidth) : '\t',
+      );
+
+      const replacementNode = templateLiteral(
+        [
+          templateElement({
+            raw: snapshot,
+          }),
+        ],
+        [],
+      );
+      args[snapshotIndex] = replacementNode;
+    },
+  });
+
+  return ast;
 };
 
 const simpleDetectParser = (filePath: Config.Path) => {