+++ packages/jest-types/src/Config.ts
@@ -45,7 +45,6 @@ export type DefaultOptions = {
       }
     | null
     | undefined;
-  cwd: Path;
   dependencyExtractor: string | null | undefined;
   errorOnDeprecated: boolean;
   expand: boolean;
@@ -89,6 +88,7 @@ export type DefaultOptions = {
   testRegex: Array<string>;
   testResultsProcessor: string | null | undefined;
   testRunner: string | null | undefined;
+  testSequencer: string;
   testURL: string;
   timers: 'real' | 'fake';
   transform:
@@ -104,6 +104,13 @@ export type DefaultOptions = {
   watch: boolean;
   watchman: boolean;
 };
+
+export type DisplayName =
+  | string
+  | {
+      name: string;
+      color: DisplayNameColor;
+    };
 
 export type InitialOptions = {
   automock?: boolean;
@@ -130,7 +137,7 @@ export type InitialOptions = {
   dependencyExtractor?: string;
   detectLeaks?: boolean;
   detectOpenHandles?: boolean;
-  displayName?: string;
+  displayName?: DisplayName;
   expand?: boolean;
   extraGlobals?: Array<string>;
   filter?: Path;
@@ -197,6 +204,7 @@ export type InitialOptions = {
   testRegex?: string | Array<string>;
   testResultsProcessor?: string | null | undefined;
   testRunner?: string;
+  testSequencer?: string;
   testURL?: string;
   timers?: 'real' | 'fake';
   transform?: {
@@ -223,6 +231,47 @@ type NotifyMode =
   | 'change'
   | 'success-change'
   | 'failure-change';
+
+/**
+ * Hard coding this until
+ * https://github.com/chalk/chalk/pull/336
+ * gets merged
+ */
+type DisplayNameColor =
+  | 'black'
+  | 'red'
+  | 'green'
+  | 'yellow'
+  | 'blue'
+  | 'magenta'
+  | 'cyan'
+  | 'white'
+  | 'gray'
+  | 'grey'
+  | 'blackBright'
+  | 'redBright'
+  | 'greenBright'
+  | 'yellowBright'
+  | 'blueBright'
+  | 'magentaBright'
+  | 'cyanBright'
+  | 'whiteBright'
+  | 'bgBlack'
+  | 'bgRed'
+  | 'bgGreen'
+  | 'bgYellow'
+  | 'bgBlue'
+  | 'bgMagenta'
+  | 'bgCyan'
+  | 'bgWhite'
+  | 'bgBlackBright'
+  | 'bgRedBright'
+  | 'bgGreenBright'
+  | 'bgYellowBright'
+  | 'bgBlueBright'
+  | 'bgMagentaBright'
+  | 'bgCyanBright'
+  | 'bgWhiteBright';
 
 type CoverageThreshold = {
   [path: string]: {
@@ -293,6 +342,7 @@ export type GlobalConfig = {
   testNamePattern: string;
   testPathPattern: string;
   testResultsProcessor: string | null | undefined;
+  testSequencer: string;
   updateSnapshot: SnapshotUpdateState;
   useStderr: boolean;
   verbose: boolean | null | undefined;
@@ -319,7 +369,7 @@ export type ProjectConfig = {
   dependencyExtractor?: string;
   detectLeaks: boolean;
   detectOpenHandles: boolean;
-  displayName: string | null | undefined;
+  displayName?: DisplayName;
   errorOnDeprecated: boolean;
   extraGlobals: Array<keyof NodeJS.Global>;
   filter: Path | null | undefined;
@@ -436,6 +486,7 @@ export type Argv = Arguments<
     testRegex: string | Array<string>;
     testResultsProcessor: string | null | undefined;
     testRunner: string;
+    testSequencer: string;
     testURL: string;
     timers: string;
     transform: string;